+++ GameServer/gameutils/LootGeneratorAtlanteanGlass.cs
@@ -58,7 +58,7 @@ namespace DOL.GS
 				DbItemTemplate atlanteanGlass = GameServer.Database.FindObjectByKey<DbItemTemplate>(m_atlanteanglass.Id_nb);
 				// ItemTemplate atlanteanGlass = new ItemTemplate(m_atlanteanglass);  Creating a new ItemTemplate throws an exception later
 
-				int killedcon = (int)player.GetConLevel(mob)+3;
+				int killedcon = player.GetConLevel(mob)+3;
 				
 				if(killedcon <= 0)
 					return loot;
