+++ GameServer/managers/RandomObjectGeneration/AtlasMobLoot.cs
@@ -39,7 +39,7 @@ namespace DOL.GS {
                     return loot;
                 }
 
-                int killedcon = (int)player.GetConLevel(mob); 
+                int killedcon = player.GetConLevel(mob); 
 
                 //grey con dont drop loot
                 if (killedcon <= -3)
