+++ GameServer/scripts/customnpc/LootGenExpOrb.cs
@@ -56,7 +56,7 @@ namespace DOL.GS
                     return loot;
                 }
                 
-                int killedcon = (int)player.GetConLevel(mob) + 3;
+                int killedcon = player.GetConLevel(mob) + 3;
 
                 if (killedcon <= 0)
                 {
